requirements.txt
# Core dependencies for Crowd Analysis
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.6.0
pandas>=1.5.0
scipy>=1.10.0
imutils>=0.5.4

# Deep learning and tracking
tensorflow>=2.12.0
keras>=2.12.0

# Machine learning
scikit-learn>=1.3.0

# Additional utilities
Pillow>=9.0.0
requests>=2.28.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Logging and configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Performance monitoring
psutil>=5.9.0

# For web dashboard (optional)
flask>=2.3.0
python-socketio>=5.8.0

# For database support (optional)
sqlalchemy>=2.0.0
